feat: coding style parser
